pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cxr-project"
version = "0.1.0"
description = "Local-first, DCC-ready chest X-ray classification scaffold for BIOSTAT 826"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
  "lightning>=2.4,<3.0",
  "matplotlib>=3.8,<4.0",
  "numpy>=1.26,<3.0",
  "pandas>=2.2,<3.0",
  "Pillow>=10.0,<12.0",
  "PyYAML>=6.0,<7.0",
  "scikit-learn>=1.5,<2.0",
  "torch>=2.4,<2.7",
  "torchmetrics>=1.4,<2.0",
  "torchvision>=0.19,<0.22",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0,<9.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]

